+++ src/idunno.Bluesky/Graph/BlueskyServer.cs
@@ -47,9 +47,6 @@ public static partial class BlueskyServer
     // https://docs.bsky.app/docs/api/app-bsky-graph-get-mutes
     private const string GetMutesEndpoint = "/xrpc/app.bsky.graph.getMutes";
 
-    // https://docs.bsky.app/docs/api/app-bsky-graph-get-relationships
-    private const string GetRelationshipsEndpoint = "/xrpc/app.bsky.graph.getRelationships";
-
     // https://docs.bsky.app/docs/api/app-bsky-graph-get-starter-pack
     private const string GetStarterPackEndpoint = "/xrpc/app.bsky.graph.getStarterPack";
 
@@ -996,97 +993,6 @@ public static partial class BlueskyServer
         else
         {
             return new AtProtoHttpResult<PagedViewReadOnlyCollection<ProfileView>>(
-                null,
-                response.StatusCode,
-                response.HttpResponseHeaders,
-                response.AtErrorDetail,
-                response.RateLimit);
-        }
-    }
-
-    // Mismatches lexicon definition - https://github.com/bluesky-social/atproto/issues/2919
-
-    /// <summary>
-    /// Enumerates public relationships between one account, and a list of other accounts.
-    /// </summary>
-    /// <param name="actor">The <see cref="Did"/> of the actor whose follows should be enumerated.</param>
-    /// <param name="others">A list of other accounts to be related back to <paramref name="actor"/>.</param>
-    /// <param name="service">The <see cref="Uri"/> of the service to retrieve the profile from.</param>
-    /// <param name="accessCredentials">The <see cref="AccessCredentials"/> used to authenticate to <paramref name="service"/>.</param>
-    /// <param name="httpClient">An <see cref="HttpClient"/> to use when making a request to the <paramref name="service"/>.</param>
-    /// <param name="onCredentialsUpdated">An <see cref="Action{T}" /> to call if the credentials in the request need updating.</param>
-    /// <param name="loggerFactory">An instance of <see cref="ILoggerFactory"/> to use to create a logger.</param>
-    /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
-    /// <returns>The task object representing the asynchronous operation.</returns>
-    /// <exception cref="ArgumentNullException">Thrown when <paramref name="actor"/>, <paramref name="others"/>, <paramref name="service"/> or <paramref name="httpClient"/> is <see langword="null"/>.</exception>
-    /// <exception cref="ArgumentOutOfRangeException">Thrown when <paramref name="others"/> is empty, or has &gt; 30 entries.</exception>
-    [UnconditionalSuppressMessage(
-        "Trimming",
-        "IL2026:Members annotated with 'RequiresUnreferencedCodeAttribute' require dynamic access otherwise can break functionality when trimming application code",
-        Justification = "All types are preserved in the JsonSerializerOptions call to Get().")]
-    [UnconditionalSuppressMessage("AOT",
-        "IL3050:Calling members annotated with 'RequiresDynamicCodeAttribute' may break functionality when AOT compiling.",
-        Justification = "All types are preserved in the JsonSerializerOptions call to Get().")]
-    public static async Task<AtProtoHttpResult<ActorRelationships>> GetRelationships(
-        Did actor,
-        ICollection<Did> others,
-        Uri service,
-        AccessCredentials? accessCredentials,
-        HttpClient httpClient,
-        Action<AtProtoCredential>? onCredentialsUpdated = null,
-        ILoggerFactory? loggerFactory = default,
-        CancellationToken cancellationToken = default)
-    {
-        ArgumentNullException.ThrowIfNull(actor);
-        ArgumentNullException.ThrowIfNull(others);
-        ArgumentNullException.ThrowIfNull(service);
-        ArgumentNullException.ThrowIfNull(httpClient);
-
-        ArgumentOutOfRangeException.ThrowIfLessThan(others.Count, 1);
-        ArgumentOutOfRangeException.ThrowIfGreaterThan(others.Count, 30);
-
-        StringBuilder queryStringBuilder = new();
-        queryStringBuilder.Append(CultureInfo.InvariantCulture, $"actor={Uri.EscapeDataString(actor.ToString())}");
-
-        foreach (AtIdentifier other in others)
-        {
-            queryStringBuilder.Append(CultureInfo.InvariantCulture, $"&others={Uri.EscapeDataString(other.ToString())}");
-        }
-
-        string queryString = queryStringBuilder.ToString();
-
-        BlueskyHttpClient<GetRelationshipsResponse> client = new(AppViewProxy, loggerFactory);
-        AtProtoHttpResult<GetRelationshipsResponse> response = await client.Get(
-            service,
-            $"{GetRelationshipsEndpoint}?{queryString}",
-            credentials: accessCredentials,
-            httpClient: httpClient,
-            jsonSerializerOptions: BlueskyJsonSerializerOptions,
-            onCredentialsUpdated: onCredentialsUpdated,
-            cancellationToken: cancellationToken).ConfigureAwait(false);
-
-        if (response.Succeeded)
-        {
-            Dictionary<Did, Relationship> returnValue = [];
-
-            foreach (RelationshipType relationshipType in response.Result.Relationships)
-            {
-                if (relationshipType is Relationship relationship)
-                {
-                    returnValue.Add(relationship.Did, relationship);
-                }
-            }
-
-            return new AtProtoHttpResult<ActorRelationships>(
-                new ActorRelationships(response.Result.Actor, returnValue),
-                response.StatusCode,
-                response.HttpResponseHeaders,
-                response.AtErrorDetail,
-                response.RateLimit);
-        }
-        else
-        {
-            return new AtProtoHttpResult<ActorRelationships>(
                 null,
                 response.StatusCode,
                 response.HttpResponseHeaders,